dashboard/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
google-cloud-bigquery>=3.14.0
google-auth>=2.25.0
pyarrow>=14.0.0
db-dtypes>=1.2.0
